refactor: modularize game screens and add types
